Restructured bokeh app scripts
