+++ dashboard/utilities/bokeh_line_plot.py
