+++ dashboard/utilities/bokeh_map_plot.py
